pkgdown/extra.css

Perform over-representation analysis

do_ora() performs over-representation analysis (ORA) using the clusterProfiler package.

Usage

do_ora(protein_list, database = c("KEGG", "GO", "Reactome"), pval_lim = 0.05)

Arguments

protein_list

A character vector containing the protein names.

database

The database to perform the ORA. It can be either "KEGG", "GO", or "Reactome".

pval_lim

The p-value threshold to consider a term as significant.

Value

A list containing the results of the ORA.

Details

The ontology option used when database = "GO" is "BP" (Biological Process).

Examples

# Perform Differential Expression Analysis
de_res <- do_limma(example_data, example_metadata, case = "AML", wide = FALSE)

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_res$de_results |>
  dplyr::filter(sig == "significant up") |>
  dplyr::pull(Assay)

# Perform ORA with GO database
do_ora(sig_up_proteins_aml, database = "GO")
#> 'select()' returned 1:1 mapping between keys and columns
#> #
#> # over-representation test
#> #
#> #...@organism 	 Homo sapiens 
#> #...@ontology 	 BP 
#> #...@keytype 	 ENTREZID 
#> #...@gene 	 chr [1:20] "566" "100" "328" "54518" "9289" "285" "9048" "181" "51129" ...
#> #...pvalues adjusted by 'BH' with cutoff <0.05 
#> #...71 enriched terms found
#> 'data.frame':	71 obs. of  9 variables:
#>  $ ID         : chr  "GO:0050900" "GO:0050926" "GO:0045785" "GO:0030336" ...
#>  $ Description: chr  "leukocyte migration" "regulation of positive chemotaxis" "positive regulation of cell adhesion" "negative regulation of cell migration" ...
#>  $ GeneRatio  : chr  "7/20" "3/20" "6/20" "5/20" ...
#>  $ BgRatio    : chr  "396/18888" "26/18888" "485/18888" "381/18888" ...
#>  $ pvalue     : num  1.03e-07 2.60e-06 7.93e-06 3.93e-05 4.72e-05 ...
#>  $ p.adjust   : num  0.000102 0.001281 0.002606 0.008413 0.008413 ...
#>  $ qvalue     : num  6.12e-05 7.69e-04 1.56e-03 5.05e-03 5.05e-03 ...
#>  $ geneID     : chr  "566/100/9048/25/2683/199/30817" "566/285/9048" "566/100/54518/9289/25/199" "100/9289/285/51742/199" ...
#>  $ Count      : int  7 3 6 5 5 3 5 4 2 5 ...
#> #...Citation
#>  T Wu, E Hu, S Xu, M Chen, P Guo, Z Dai, T Feng, L Zhou, W Tang, L Zhan, X Fu, S Liu, X Bo, and G Yu.
#>  clusterProfiler 4.0: A universal enrichment tool for interpreting omics data.
#>  The Innovation. 2021, 2(3):100141 
#>