Differential expression analysis with t-test

do_ttest_de() performs differential expression analysis using t-test. It separates the data in case-control groups, checks for data normality and perform a t-test or Wilcoxon test respectively. It also performs p value FDR adjustment.

Usage

do_ttest_de(
  long_data,
  disease,
  assays,
  normality_res,
  only_female = NULL,
  only_male = NULL,
  pval_lim = 0.05,
  logfc_lim = 0
)

Arguments

long_data

A tibble with the Olink data in long format and the Sex column from metadata.

disease

The disease of interest.

assays

The assays to run the differential expression analysis on.

only_female

The female specific diseases. Default is NULL.

only_male

The male specific diseases. Default is NULL.

pval_lim

The p-value limit of significance. Default is 0.05.

logfc_lim

The logFC limit of significance. Default is 0.

Value

A tibble with the differential expression results.