pkgdown/extra.css

Create and save wide and join dataframes

generate_df() creates wide and join dataframes from the long data and metadata. It saves them in the data/processed/data_metadata directory in RDA format.

Usage

generate_df(
  long_data,
  metadata = NULL,
  join = TRUE,
  metadata_cols = c("DAid", "Disease", "Sex", "Age", "BMI"),
  save = TRUE
)

Arguments

long_data

The long data.

metadata

The metadata.

join

If TRUE, the dataframes are joined with metadata.

metadata_cols

The metadata columns to join with data.

save

If TRUE, the dataframes are saved.

Value

A list containing the following elements:

wide_data: The wide data.

join_data: The joined data with metadata. If join is FALSE, this is not returned.

Details

It will filter out rows with missing values in Disease, as a results of mismatching between data and metadata

Examples

generate_df(example_data, example_metadata, save = FALSE)
#> $wide_data
#> # A tibble: 586 × 101
#>    DAid    AARSD1   ABL1  ACAA1    ACAN    ACE2  ACOX1   ACP5    ACP6  ACTA2
#>    <chr>    <dbl>  <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 DA00001   3.39  2.76   1.71   0.0333  1.76   -0.919 1.54    2.15    2.81 
#>  2 DA00002   1.42  1.25  -0.816 -0.459   0.826  -0.902 0.647   1.30    0.798
#>  3 DA00003  NA    NA     NA      0.989  NA       0.330 1.37   NA      NA    
#>  4 DA00004   3.41  3.38   1.69  NA       1.52   NA     0.841   0.582   1.70 
#>  5 DA00005   5.01  5.05   0.128  0.401  -0.933  -0.584 0.0265  1.16    2.73 
#>  6 DA00006   6.83  1.18  -1.74  -0.156   1.53   -0.721 0.620   0.527   0.772
#>  7 DA00007  NA    NA      3.96   0.682   3.14    2.62  1.47    2.25    2.01 
#>  8 DA00008   2.78  0.812 -0.552  0.982  -0.101  -0.304 0.376  -0.826   1.52 
#>  9 DA00009   4.39  3.34  -0.452 -0.868   0.395   1.71  1.49   -0.0285  0.200
#> 10 DA00010   1.83  1.21  -0.912 -1.04   -0.0918 -0.304 1.69    0.0920  2.04 
#> # ℹ 576 more rows
#> # ℹ 91 more variables: ACTN4 <dbl>, ACY1 <dbl>, ADA <dbl>, ADA2 <dbl>,
#> #   ADAM15 <dbl>, ADAM23 <dbl>, ADAM8 <dbl>, ADAMTS13 <dbl>, ADAMTS15 <dbl>,
#> #   ADAMTS16 <dbl>, ADAMTS8 <dbl>, ADCYAP1R1 <dbl>, ADGRE2 <dbl>, ADGRE5 <dbl>,
#> #   ADGRG1 <dbl>, ADGRG2 <dbl>, ADH4 <dbl>, ADM <dbl>, AGER <dbl>, AGR2 <dbl>,
#> #   AGR3 <dbl>, AGRN <dbl>, AGRP <dbl>, AGXT <dbl>, AHCY <dbl>, AHSP <dbl>,
#> #   AIF1 <dbl>, AIFM1 <dbl>, AK1 <dbl>, AKR1B1 <dbl>, AKR1C4 <dbl>, …
#> 
#> $join_data
#> # A tibble: 586 × 105
#>    DAid    Disease Sex     Age   BMI AARSD1   ABL1  ACAA1    ACAN    ACE2  ACOX1
#>    <chr>   <chr>   <chr> <dbl> <dbl>  <dbl>  <dbl>  <dbl>   <dbl>   <dbl>  <dbl>
#>  1 DA00001 AML     F        42  22.7   3.39  2.76   1.71   0.0333  1.76   -0.919
#>  2 DA00002 AML     M        69  33.1   1.42  1.25  -0.816 -0.459   0.826  -0.902
#>  3 DA00003 AML     F        61  26.2  NA    NA     NA      0.989  NA       0.330
#>  4 DA00004 AML     M        54  28.1   3.41  3.38   1.69  NA       1.52   NA    
#>  5 DA00005 AML     F        57  21.4   5.01  5.05   0.128  0.401  -0.933  -0.584
#>  6 DA00006 AML     M        86  33.9   6.83  1.18  -1.74  -0.156   1.53   -0.721
#>  7 DA00007 AML     F        85  28.7  NA    NA      3.96   0.682   3.14    2.62 
#>  8 DA00008 AML     F        88  32.6   2.78  0.812 -0.552  0.982  -0.101  -0.304
#>  9 DA00009 AML     M        80  26.1   4.39  3.34  -0.452 -0.868   0.395   1.71 
#> 10 DA00010 AML     M        48  33.8   1.83  1.21  -0.912 -1.04   -0.0918 -0.304
#> # ℹ 576 more rows
#> # ℹ 94 more variables: ACP5 <dbl>, ACP6 <dbl>, ACTA2 <dbl>, ACTN4 <dbl>,
#> #   ACY1 <dbl>, ADA <dbl>, ADA2 <dbl>, ADAM15 <dbl>, ADAM23 <dbl>, ADAM8 <dbl>,
#> #   ADAMTS13 <dbl>, ADAMTS15 <dbl>, ADAMTS16 <dbl>, ADAMTS8 <dbl>,
#> #   ADCYAP1R1 <dbl>, ADGRE2 <dbl>, ADGRE5 <dbl>, ADGRG1 <dbl>, ADGRG2 <dbl>,
#> #   ADH4 <dbl>, ADM <dbl>, AGER <dbl>, AGR2 <dbl>, AGR3 <dbl>, AGRN <dbl>,
#> #   AGRP <dbl>, AGXT <dbl>, AHCY <dbl>, AHSP <dbl>, AIF1 <dbl>, AIFM1 <dbl>, …
#>