pkgdown/extra.css

Filter data and metadata by sex

hd_filter_by_sex() filters the data and metadata by the metadata Sex variable and a specified value. It can be used in cases of sex specific diseases before running differential expression or classification analysis.

Usage

hd_filter_by_sex(dat, metadata = NULL, variable = "Sex", sex)

Arguments

dat

An HDAnalyzeR object or a dataset in wide format and sample ID as its first column.

metadata

A dataset containing the metadata information with the sample ID as the first column. If a HDAnalyzeR object is provided, this parameter is not needed.

variable

The name of the variable in the metadata contain information for sex. Default is "Sex".

sex

The value of the sex variable to filter by.

Value

A list containing the filtered data and metadata.

Examples

# Create the HDAnalyzeR object providing the data and metadata
hd_object <- hd_initialize(example_data, example_metadata)

hd_filter_by_sex(hd_object, variable = "Sex", sex = "F")
#> $data
#> # A tibble: 366 × 101
#>    DAid  AARSD1   ABL1  ACAA1    ACAN   ACE2   ACOX1    ACP5   ACP6 ACTA2  ACTN4
#>    <chr>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl> <dbl>  <dbl>
#>  1 DA00…   3.39  2.76   1.71   0.0333  1.76  -0.919   1.54    2.15   2.81  0.742
#>  2 DA00…  NA    NA     NA      0.989  NA      0.330   1.37   NA     NA    NA    
#>  3 DA00…   5.01  5.05   0.128  0.401  -0.933 -0.584   0.0265  1.16   2.73  0.350
#>  4 DA00…  NA    NA      3.96   0.682   3.14   2.62    1.47    2.25   2.01  0.170
#>  5 DA00…   2.78  0.812 -0.552  0.982  -0.101 -0.304   0.376  -0.826  1.52 -0.597
#>  6 DA00…   3.48  4.96   3.50  -0.338   4.48   1.26    2.18    1.62   1.79  0.233
#>  7 DA00…   4.31  0.710 -1.44  -0.218  -0.469 -0.361  -0.0714 -1.30   2.86  0.129
#>  8 DA00…   1.79  1.36   0.106 -0.372   3.40  -1.19    1.77    1.07   2.00  0.980
#>  9 DA00…   3.59  3.38   1.79  -0.303   1.59   0.604   1.71   -0.837  1.65  1.13 
#> 10 DA00…   1.80  1.70   2.77  -1.04    1.33  -0.0247  1.02    0.112  2.58  1.14 
#> # ℹ 356 more rows
#> # ℹ 90 more variables: ACY1 <dbl>, ADA <dbl>, ADA2 <dbl>, ADAM15 <dbl>,
#> #   ADAM23 <dbl>, ADAM8 <dbl>, ADAMTS13 <dbl>, ADAMTS15 <dbl>, ADAMTS16 <dbl>,
#> #   ADAMTS8 <dbl>, ADCYAP1R1 <dbl>, ADGRE2 <dbl>, ADGRE5 <dbl>, ADGRG1 <dbl>,
#> #   ADGRG2 <dbl>, ADH4 <dbl>, ADM <dbl>, AGER <dbl>, AGR2 <dbl>, AGR3 <dbl>,
#> #   AGRN <dbl>, AGRP <dbl>, AGXT <dbl>, AHCY <dbl>, AHSP <dbl>, AIF1 <dbl>,
#> #   AIFM1 <dbl>, AK1 <dbl>, AKR1B1 <dbl>, AKR1C4 <dbl>, AKT1S1 <dbl>, …
#> 
#> $metadata
#> # A tibble: 366 × 9
#>    DAid    Sample     Disease Stage Grade Sex     Age   BMI Cohort
#>    <chr>   <chr>      <chr>   <chr> <chr> <chr> <dbl> <dbl> <chr> 
#>  1 DA00001 AML_syn_1  AML     2     NA    F        42  22.7 UCAN  
#>  2 DA00003 AML_syn_3  AML     2     NA    F        61  26.2 UCAN  
#>  3 DA00005 AML_syn_5  AML     2     NA    F        57  21.4 UCAN  
#>  4 DA00007 AML_syn_7  AML     1     NA    F        85  28.7 UCAN  
#>  5 DA00008 AML_syn_8  AML     3     NA    F        88  32.6 UCAN  
#>  6 DA00011 AML_syn_11 AML     3     NA    F        54  34.7 UCAN  
#>  7 DA00012 AML_syn_12 AML     3     NA    F        78  21.4 UCAN  
#>  8 DA00016 AML_syn_16 AML     3     NA    F        78  25.4 UCAN  
#>  9 DA00019 AML_syn_19 AML     1     NA    F        81  22.9 UCAN  
#> 10 DA00020 AML_syn_20 AML     3     NA    F        65  24.1 UCAN  
#> # ℹ 356 more rows
#> 
#> $sample_id
#> [1] "DAid"
#> 
#> $var_name
#> [1] "Assay"
#> 
#> $value_name
#> [1] "NPX"
#> 
#> attr(,"class")
#> [1] "HDAnalyzeR"