pkgdown/extra.css

Run over-representation analysis

hd_run_ora() performs over-representation analysis (ORA) using the clusterProfiler package.

Usage

hd_run_ora(
  gene_list,
  database = c("GO", "Reactome", "KEGG"),
  ontology = c("BP", "CC", "MF", "ALL"),
  background = NULL,
  pval_lim_enrichment = 0.05
)

Arguments

gene_list

A character vector containing the gene names. These can be differentially expressed proteins or selected protein features from classification models.

database

The database to perform the ORA. It can be either "GO", "KEGG", or "Reactome".

ontology

The ontology to use when database = "GO". It can be "BP" (Biological Process), "CC" (Cellular Component), "MF" (Molecular Function), or "ALL". In the case of KEGG and Reactome, this parameter is ignored.

background

A character vector containing the background genes.

pval_lim_enrichment

The p-value threshold to consider a term as significant in the enrichment analysis.

Value

A list containing the results of the ORA.

Details

qvalueCutoff was set to be 4 times the pvalueCutoff.

Examples

# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_run_de_limma(hd_object, case = "AML")

# Extract the up-regulated proteins for AML
sig_up_proteins_aml <- de_results$de_res |>
  dplyr::filter(adj.P.Val < 0.05 & logFC > 0) |>
  dplyr::pull(Feature)

# Perform ORA with `GO` database and `BP` ontology
hd_run_ora(sig_up_proteins_aml, database = "GO", ontology = "BP")
#> 'select()' returned 1:1 mapping between keys and columns
#> $gene_list
#>  [1] "100"   "566"   "328"   "54518" "9289"  "285"   "9048"  "181"   "51129"
#> [10] "191"   "25"    "59"    "2683"  "51742" "9296"  "267"   "59272" "976"  
#> [19] "199"   "405"   "30817" "1386" 
#> 
#> $enrichment
#> #
#> # over-representation test
#> #
#> #...@organism 	 Homo sapiens 
#> #...@ontology 	 BP 
#> #...@keytype 	 ENTREZID 
#> #...@gene 	 chr [1:22] "100" "566" "328" "54518" "9289" "285" "9048" "181" "51129" ...
#> #...pvalues adjusted by 'BH' with cutoff <0.05 
#> #...83 enriched terms found
#> 'data.frame':	83 obs. of  12 variables:
#>  $ ID            : chr  "GO:0050900" "GO:0050926" "GO:0071674" "GO:0045785" ...
#>  $ Description   : chr  "leukocyte migration" "regulation of positive chemotaxis" "mononuclear cell migration" "positive regulation of cell adhesion" ...
#>  $ GeneRatio     : chr  "7/22" "3/22" "5/22" "6/22" ...
#>  $ BgRatio       : chr  "404/18986" "27/18986" "231/18986" "499/18986" ...
#>  $ RichFactor    : num  0.0173 0.1111 0.0216 0.012 0.0158 ...
#>  $ FoldEnrichment: num  15 95.9 18.7 10.4 13.6 ...
#>  $ zScore        : num  9.66 16.81 9.21 7.23 7.71 ...
#>  $ pvalue        : num  2.43e-07 3.88e-06 5.68e-06 1.67e-05 2.62e-05 ...
#>  $ p.adjust      : num  0.000285 0.002226 0.002226 0.004903 0.006166 ...
#>  $ qvalue        : num  0.000173 0.001349 0.001349 0.002971 0.003737 ...
#>  $ geneID        : chr  "100/566/9048/25/2683/199/30817" "566/285/9048" "566/9048/25/2683/199" "100/566/54518/9289/25/199" ...
#>  $ Count         : int  7 3 5 6 5 5 3 5 5 3 ...
#> #...Citation
#> S Xu, E Hu, Y Cai, Z Xie, X Luo, L Zhan, W Tang, Q Wang, B Liu, R Wang, W Xie, T Wu, L Xie, G Yu. Using clusterProfiler to characterize multiomics data. Nature Protocols. 2024, doi:10.1038/s41596-024-01020-z 
#> 
#> 
#> attr(,"class")
#> [1] "hd_enrichment"