pkgdown/extra.css

Automated PubMed literature search

literature_search() searches for articles for protein-disease pairs in PubMed. A list of proteins and diseases is provided as input. The function retrieves the articles for each protein-disease pair. The input should be in the correct format, a list with diseases as names and protein vectors associated with each disease as elements (see examples).

Usage

literature_search(prot_dis_list, max_articles = 10)

Arguments

prot_dis_list

A list of proteins and diseases. The names of the list are the diseases and the elements are vectors of proteins.

max_articles

The maximum number of articles to retrieve for each protein-disease pair. Default is 10.

Value

A list of tibbles. Each tibble contains the articles found for a protein-disease pair.

Details

The disease and gene names should be correct in order for the query to be successful. For example AML should be written as "acute myeloid leukemia".

Examples

# Prepare the list of protein-disease pairs
prot_dis_list <- list("acute myeloid leukemia" = c("FLT3", "EPO"),
                      "chronic lymphocytic leukemia" = c("PARP1", "FCER2"))

# Run the literature search
lit_search_results <- literature_search(prot_dis_list, max_articles = 1)
#> Searching for articles on FLT3 and acute myeloid leukemia
#> Searching for articles on EPO and acute myeloid leukemia
#> Searching for articles on PARP1 and chronic lymphocytic leukemia
#> Searching for articles on FCER2 and chronic lymphocytic leukemia

# Results for FLT3 in acute myeloid leukemia
lit_search_results[["acute myeloid leukemia"]][["FLT3"]]
#>       pmid        First_author year                   journal
#> 1 39073493 Arwanih, Elly Yanah 2024 Molecular biology reports
#>                                                                                                                           title
#> 1 Identification of a novel mutation of the FLT3 gene located on the juxtamembrane domain from acute myeloid leukemia patients.