Plot sample data points in a two-dimensional plane.

plot_dim_reduction() plots the sample data points in a two-dimensional plane. The points can be plotted in the PC1/PC2 or UMAP1/UMAP2 space.

Usage

plot_dim_reduction(res, x, y, metadata, color, palette)

Arguments

res

A tibble with the results of the dimensionality reduction analysis.

x

The name of the column in res that contains the x-axis values.

y

The name of the column in res that contains the y-axis values.

metadata

A tibble with metadata information to be added to the plot. It will be used for coloring the points.

color

The name of the column in metadata that contains the variable to be used to plot the points color.

palette

The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes().

Value

A ggplot object