pkgdown/extra.css

Plot feature variable importance

plot_var_imp() collects the features and their model importance. It scales their importance and plots it against them.

Usage

plot_var_imp(
  finalfit_res,
  case,
  accuracy,
  sensitivity,
  specificity,
  auc,
  mixture,
  palette = NULL,
  vline = TRUE,
  subtitle = c("accuracy", "sensitivity", "specificity", "auc", "features",
    "top-features", "mixture"),
  yaxis_names = FALSE
)

Arguments

finalfit_res

Results from finalfit().

case

Case to predict.

accuracy

Accuracy of the model.

sensitivity

Sensitivity of the model.

specificity

Specificity of the model.

auc

AUC of the model.

mixture

Mixture of lasso and ridge regularization.

palette

The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is NULL.

vline

Whether to add a vertical line at 50% importance. Default is TRUE.

subtitle

Vector of subtitle elements to include in the plot.

yaxis_names

Whether to add y-axis names to the plot. Default is FALSE.

Value

A list with two elements:

features: A tibble with features and their model importance.

var_imp_plot: Variable importance plot.