pkgdown/extra.css

Create volcano plots

plot_volcano() creates volcano plots for the differential expression results. It colors and labels the top up and down regulated proteins.

Usage

plot_volcano(
  de_result,
  pval_lim = 0.05,
  logfc_lim = 0,
  top_up_prot = 40,
  top_down_prot = 10,
  palette = "diff_exp",
  title = NULL,
  report_nproteins = TRUE,
  user_defined_proteins = NULL,
  subtitle = NULL
)

Arguments

de_result

The differential expression results.

pval_lim

The p-value limit for significance. Default is 0.05.

logfc_lim

The logFC limit for significance. Default is 0.

top_up_prot

The number of top up regulated proteins to label on the plot. Default is 40.

top_down_prot

The number of top down regulated proteins to label on the plot. Default is 10.

palette

The color palette for the plot. If it is a character, it should be one of the palettes from get_hpa_palettes(). Default is "diff_exp".

title

The title of the plot or NULL for no title.

report_nproteins

If the number of significant proteins should be reported in the subtitle. Default is TRUE.

subtitle

The subtitle of the plot or NULL for no subtitle.

Value

A ggplot object with the volcano plot.