pkgdown/extra.css

Print the summary of the quality control results

print_summary() prints the summary of the quality control results of the input dataset. It includes the number of samples and variables, the counts of each class, the percentage of NAs in each column and row, the normality test results, the protein-protein correlations above a certain threshold, and the correlation heatmap.

Usage

print_summary(
  sample_n,
  var_n,
  class_summary,
  na_percentage_col,
  na_percentage_row,
  normality_results = FALSE,
  cor_results = FALSE,
  heatmap = FALSE,
  threshold = FALSE
)

Arguments

sample_n

The number of samples.

var_n

The number of variables.

class_summary

A table with the counts of each class in the dataframe.

na_percentage_col

A tibble with the column names and the percentage of NAs in each column.

na_percentage_row

A tibble with the DAids and the percentage of NAs in each row.

normality_results

A tibble with the protein names, p-values, adjusted p-values, and normality status.

cor_results

A tibble with the filtered protein pairs and their correlation values.

heatmap

A heatmap of protein-protein correlations.

threshold

The reporting protein-protein correlation threshold.