pkgdown/extra.css

Extract model features and plot variable importance

variable_imp() calculates the variable importance of the model and plots the top features. It also generates a title for the plot based on the model metrics and the mixture parameter.

Usage

variable_imp(
  dat,
  variable = "Disease",
  case,
  mixture = NULL,
  palette = NULL,
  y_labels = FALSE,
  title = c("accuracy", "sensitivity", "specificity", "auc", "features", "top-features"),
  verbose = TRUE,
  seed = 123
)

Arguments

dat

An hd_model object coming from evaluate_model().

variable

The variable to predict. Default is "Disease".

case

The case class.

mixture

The mixture parameter for the elastic net. If NULL it will be tuned. Default is NULL.

palette

The color palette for the classes. If it is a character, it should be one of the palettes from hd_palettes(). Default is NULL.

title

Vector of title elements to include in the plot.

seed

Seed for reproducibility. Default is 123.

Value

A model object containing the features and the variable importance plot.