pkgdown/extra.css

Gene set enrichment analysis

hd_gsea() performs gene set enrichment analysis (GSEA) using the clusterProfiler package.

Usage

hd_gsea(
  de_results,
  database = c("GO", "Reactome", "KEGG"),
  ontology = c("BP", "CC", "MF", "ALL"),
  ranked_by = "logFC",
  pval_lim = 0.05
)

Arguments

de_results

An hd_de object from hd_de_limma() or hd_de_ttest() or a tibble containing the results of a differential expression analysis.

database

The database to perform the ORA. It can be either "GO", "KEGG", or "Reactome".

ontology

The ontology to use when database = "GO". It can be "BP" (Biological Process), "CC" (Cellular Component), "MF" (Molecular Function), or "ALL". In the case of KEGG and Reactome, this parameter is ignored.

ranked_by

The variable to rank the proteins. It can be "logFC", "both" which is the product of logFC and -log(adj.P.Val) or a custom sorting variable. It should be however a column in the DE results tibble (de_results argument).

pval_lim

The p-value threshold to consider a term as significant in the enrichment analysis. Default is 0.05.

Value

A list containing the results of the GSEA.

Details

To perform the GSEA, clusterProfiler package is used. For more information, please refer to the clusterProfiler documentation.

If you want to learn more about GSEA, please refer to the following publications:

Chicco D, Agapito G. Nine quick tips for pathway enrichment analysis. PLoS Comput Biol. 2022 Aug 11;18(8):e1010348. doi: 10.1371/journal.pcbi.1010348. PMID: 35951505; PMCID: PMC9371296. https://pmc.ncbi.nlm.nih.gov/articles/PMC9371296/

https://yulab-smu.top/biomedical-knowledge-mining-book/enrichment-overview.html#gsea-algorithm

Examples

# Initialize an HDAnalyzeR object
hd_object <- hd_initialize(example_data, example_metadata)

# Run differential expression analysis for AML vs all others
de_results <- hd_de_limma(hd_object, case = "AML")

# Run GSEA with GO database
hd_gsea(de_results,
        database = "GO",
        ontology = "BP",
        ranked_by = "logFC",
        pval_lim = 0.9)
#> 
#> 
#> 'select()' returned 1:1 mapping between keys and columns
#> using 'fgsea' for GSEA analysis, please cite Korotkevich et al (2019).
#> preparing geneSet collections...
#> GSEA analysis...
#> leading edge analysis...
#> done...
#> $gene_list
#>          328          566          100         9289        54518         9048 
#>  1.559583713  1.532943471  1.462382647  1.231859388  1.173013982  0.828793582 
#>          285          181           25          191        51129         9296 
#>  0.772836518  0.756171703  0.746830743  0.734041132  0.522682913  0.519848239 
#>           59        59272        51742        25814          199          267 
#>  0.504682910  0.466035852  0.458170283  0.451765263  0.432416150  0.428812794 
#>         2683         9938          976        30817         1386          405 
#>  0.421745793  0.397319743  0.365431820  0.359446007  0.338581655  0.337841863 
#>        23452        10159        51816          374          410          133 
#>  0.292444281  0.271377450  0.253959306  0.242008838  0.226357122  0.199535962 
#>       170689        51327          411        84335          214       375790 
#>  0.177164675  0.175706968  0.173203848  0.170442918  0.162775605  0.159570184 
#>         8751           95       129138          308        80755          558 
#>  0.123946384  0.122930955  0.113185504  0.108778505  0.106165339  0.093214100 
#>         9068          283          218          258          189        81693 
#>  0.092176270  0.086140961  0.081630236  0.059789250  0.045913185  0.043794218 
#>          290           54           81         8745           51         1109 
#>  0.026847324  0.019847155 -0.003178737 -0.009723351 -0.009971140 -0.016368666 
#>          231          177          383          117          392          350 
#> -0.018310077 -0.038228906 -0.039306455 -0.043239615 -0.054674439 -0.076035424 
#>        51382          539          333          419       155465         9131 
#> -0.097147761 -0.098498094 -0.104025127 -0.104158209 -0.104553945 -0.110089916 
#>          127         8312          101          259           30        27329 
#> -0.129895411 -0.135089674 -0.138922933 -0.160969394 -0.161931907 -0.183584355 
#>         8639          216       347902          279        11199           26 
#> -0.194111783 -0.195512297 -0.200816339 -0.215411972 -0.244774741 -0.245691583 
#>        10218       170690        51205        11093        10000          475 
#> -0.272748339 -0.298014216 -0.308354203 -0.314105581 -0.333881970 -0.337183300 
#>        11095          307        93974        10149          280          176 
#> -0.339134804 -0.342512836 -0.356798140 -0.375853514 -0.376116879 -0.383419881 
#>          311        55937          203          306       115201        10551 
#> -0.404836340 -0.440825951 -0.481572885 -0.619682407 -0.629226223 -0.704655666 
#>        23365          351          250          284 
#> -0.806377522 -0.822744636 -1.036735621 -1.696033915 
#> 
#> $enrichment
#> #
#> # Gene Set Enrichment Analysis
#> #
#> #...@organism 	 Homo sapiens 
#> #...@setType 	 BP 
#> #...@keytype 	 ENTREZID 
#> #...@geneList 	 Named num [1:100] 1.56 1.53 1.46 1.23 1.17 ...
#>  - attr(*, "names")= chr [1:100] "328" "566" "100" "9289" ...
#> #...nPerm 	 
#> #...pvalues adjusted by 'BH' with cutoff <0.9 
#> #...238 enriched terms found
#> 'data.frame':	238 obs. of  11 variables:
#>  $ ID             : chr  "GO:0015031" "GO:0045184" "GO:0071495" "GO:0048518" ...
#>  $ Description    : chr  "protein transport" "establishment of protein localization" "cellular response to endogenous stimulus" "positive regulation of biological process" ...
#>  $ setSize        : int  10 11 13 42 76 71 20 16 16 11 ...
#>  $ enrichmentScore: num  -0.769 -0.706 0.738 0.551 0.552 ...
#>  $ NES            : num  -1.87 -1.78 1.76 1.64 1.62 ...
#>  $ pvalue         : num  0.00191 0.00693 0.00487 0.00657 0.00506 ...
#>  $ p.adjust       : num  0.282 0.282 0.282 0.282 0.282 ...
#>  $ qvalue         : num  0.267 0.267 0.267 0.267 0.267 ...
#>  $ rank           : num  7 15 24 19 28 28 15 7 7 7 ...
#>  $ leading_edge   : chr  "tags=30%, list=7%, signal=31%" "tags=36%, list=15%, signal=35%" "tags=62%, list=24%, signal=54%" "tags=36%, list=19%, signal=50%" ...
#>  $ core_enrichment: chr  "115201/351/284" "93974/115201/351/284" "9289/285/181/25/59/51742/1386/405" "328/566/100/9289/54518/9048/285/181/25/51129/59/59272/51742/199/2683" ...
#> #...Citation
#> S Xu, E Hu, Y Cai, Z Xie, X Luo, L Zhan, W Tang, Q Wang, B Liu, R Wang, W Xie, T Wu, L Xie, G Yu. Using clusterProfiler to characterize multiomics data. Nature Protocols. 2024, 19(11):3292-3320 
#> 
#> 
#> attr(,"class")
#> [1] "hd_enrichment"
# Remember that the data is artificial, this is why we use an absurdly high p-value cutoff

# Run GSEA with different ranking variable
hd_gsea(de_results,
        database = "GO",
        ontology = "BP",
        ranked_by = "both",
        pval_lim = 0.9)
#> 'select()' returned 1:1 mapping between keys and columns
#> using 'fgsea' for GSEA analysis, please cite Korotkevich et al (2019).
#> preparing geneSet collections...
#> GSEA analysis...
#> Warning: There are ties in the preranked stats (22% of the list).
#> The order of those tied genes will be arbitrary, which may produce unexpected results.
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> leading edge analysis...
#> done...
#> $gene_list
#>           81           51         8745          231         1109           54 
#> 9.762289e-01 9.553779e-01 9.539886e-01 9.439470e-01 9.363257e-01 9.122193e-01 
#>          189          290          383          177        81693         9131 
#> 8.749648e-01 8.749648e-01 8.552561e-01 8.552561e-01 8.440693e-01 8.232858e-01 
#>          117          392          218          258          333          350 
#> 8.037043e-01 7.787213e-01 6.919034e-01 6.919034e-01 6.508476e-01 6.508476e-01 
#>         8312        80755          127          283           95         9068 
#> 6.496255e-01 6.492551e-01 6.492551e-01 6.140926e-01 6.080472e-01 6.078419e-01 
#>        51327           30          539       155465        51382       129138 
#> 5.843261e-01 5.835408e-01 5.835408e-01 5.835408e-01 5.809570e-01 5.809570e-01 
#>          558          419          308        84335          411          101 
#> 5.809570e-01 5.435734e-01 5.403000e-01 5.321584e-01 5.231048e-01 4.741485e-01 
#>         8751       375790        93974       170689        11199          216 
#> 4.509795e-01 3.759502e-01 3.734635e-01 3.567013e-01 2.847430e-01 2.816466e-01 
#>          214          259       347902        27329         8639          279 
#> 2.795845e-01 2.525653e-01 2.302575e-01 2.275316e-01 2.275316e-01 2.275316e-01 
#>          133          374          307        51816           26          410 
#> 2.113606e-01 1.938142e-01 1.671439e-01 1.593498e-01 1.517575e-01 1.513815e-01 
#>          475        23452        51205        10159        11095        10218 
#> 1.513815e-01 1.152428e-01 7.837423e-02 6.809229e-02 6.666420e-02 6.666420e-02 
#>        10000         9938        25814       170690         1386          311 
#> 6.236680e-02 5.312927e-02 5.312927e-02 3.900195e-02 3.900195e-02 3.900195e-02 
#>        30817        11093          199          405          203          976 
#> 2.777020e-02 2.050877e-02 2.050877e-02 2.050877e-02 2.011191e-02 1.879857e-02 
#>          280        59272        10551          306          267        10149 
#> 1.879857e-02 1.845092e-02 1.459705e-02 1.058872e-02 1.040728e-02 9.562547e-03 
#>        23365          176         9296         2683        51742           59 
#> 6.027175e-03 5.315769e-03 4.772619e-03 4.051888e-03 4.051888e-03 3.692718e-03 
#>           25          191        55937          250        51129       115201 
#> 3.487500e-03 3.164572e-03 3.164572e-03 6.657176e-04 6.034845e-04 4.720349e-04 
#>          351          181         9048          285         9289          284 
#> 1.348131e-05 1.348131e-05 8.685788e-08 3.857309e-08 2.920315e-08 2.480060e-10 
#>        54518          328          566          100 
#> 3.687720e-11 4.668228e-13 6.621823e-17 3.293393e-19 
#> 
#> $enrichment
#> #
#> # Gene Set Enrichment Analysis
#> #
#> #...@organism 	 Homo sapiens 
#> #...@setType 	 BP 
#> #...@keytype 	 ENTREZID 
#> #...@geneList 	 Named num [1:100] 0.976 0.955 0.954 0.944 0.936 ...
#>  - attr(*, "names")= chr [1:100] "81" "51" "8745" "231" ...
#> #...nPerm 	 
#> #...pvalues adjusted by 'BH' with cutoff <0.9 
#> #...180 enriched terms found
#> 'data.frame':	180 obs. of  11 variables:
#>  $ ID             : chr  "GO:0003008" "GO:0080134" "GO:0090304" "GO:0044093" ...
#>  $ Description    : chr  "system process" "regulation of response to stress" "nucleic acid metabolic process" "positive regulation of molecular function" ...
#>  $ setSize        : int  15 13 16 10 12 40 55 25 13 15 ...
#>  $ enrichmentScore: num  0.726 0.733 0.715 0.747 0.729 ...
#>  $ NES            : num  1.31 1.3 1.3 1.29 1.28 ...
#>  $ pvalue         : num  0.0209 0.036 0.028 0.0533 0.0511 ...
#>  $ p.adjust       : num  0.857 0.857 0.857 0.857 0.857 ...
#>  $ qvalue         : num  0.857 0.857 0.857 0.857 0.857 ...
#>  $ rank           : int  19 18 22 19 10 25 39 39 18 22 ...
#>  $ leading_edge   : chr  "tags=27%, list=19%, signal=25%" "tags=38%, list=18%, signal=36%" "tags=44%, list=22%, signal=41%" "tags=50%, list=19%, signal=45%" ...
#>  $ core_enrichment: chr  "231/177/81693/8312" "54/383/177/117/350" "81/177/9131/333/8312/80755/283" "81/177/9131/350/8312" ...
#> #...Citation
#> S Xu, E Hu, Y Cai, Z Xie, X Luo, L Zhan, W Tang, Q Wang, B Liu, R Wang, W Xie, T Wu, L Xie, G Yu. Using clusterProfiler to characterize multiomics data. Nature Protocols. 2024, 19(11):3292-3320 
#> 
#> 
#> attr(,"class")
#> [1] "hd_enrichment"